sr(api): move create-or-update-training-trigger in devcomp
